requirements.txt
mysql-connector-python
requests
bs4
sqlalchemy = 2.0.25